finished initial impl of batcher sort

+++ odd-even.cu
@@ -38,7 +38,7 @@ __global__ void smallOddEvenSort(float vec[], size_t size) {
 
 // sort a partition of the global vector on each block using bitonic sort
 __global__ void localBitonicSort(float vec[], size_t size) {
-    int idx = blockDim.x * blockIdx.x + threadIdx.x;
+    size_t idx = blockDim.x * blockIdx.x + threadIdx.x;
     if (idx >= size) {
         return;
     }
@@ -64,7 +64,7 @@ __global__ void localBitonicSort(float vec[], size_t size) {
 
 // bitonic merge two sorted subarrays/partitions into one subarray/partition
 __global__ void bitonicMerge(float vec[], size_t size, bool even) {
-    int idx = blockDim.x * blockIdx.x + threadIdx.x;
+    size_t idx = blockDim.x * blockIdx.x + threadIdx.x;
     if (idx >= size || (!even && idx + NUM_THREADS >= size)) {
         return;
     }
